Fix IdVersionTuple id to long

+++ src/main/java/org/opensrp/domain/IdVersionTuple.java
@@ -3,19 +3,19 @@ package org.opensrp.domain;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
 public class IdVersionTuple {
-    public int id;
+    public long id;
     public String version;
 
-    public IdVersionTuple(int id, @NonNull String version) {
+    public IdVersionTuple(long id, @NonNull String version) {
         this.id = id;
         this.version = version;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
